Consolidate Info Loading
Instead of individual loaders for BaseLinker and Refiner, we
consolidate into a single loader.
